+++ ivy/projects/lib-demo-code/src/lib/on-changes.component.ts
@@ -7,16 +7,24 @@ import { input$, OnChanges$ } from '@pdtec/ngx-observable-lifecycle';
 })
 export class OnChangesComponent extends OnChanges$ {
   @Input()
-  public value: number | undefined = 5;
+  public value1: number | undefined;
+
+  @Input()
+  public value2: number | undefined;
 
   public valueFromObservable: number | undefined;
 
   constructor() {
     super();
 
-    input$(this, 'value')
+    input$(this, 'value1')
       .subscribe(value => {
         this.valueFromObservable = value;
       });
+
+    this.inputs$('value1', 'value2').subscribe(x => {
+      console.log(`x.value`, x.value1);
+      console.log(`x.value2`, x.value2);
+    });
   }
 }