+++ ivy/projects/lib/src/lib/hooks/on-changes.ts
@@ -1,6 +1,7 @@
 import { OnChanges, SimpleChanges } from '@angular/core';
 import { Observable, ReplaySubject } from 'rxjs';
 import { OnDestroy$ } from './on-destroy';
+import { input$, inputs$Impl } from '../operators/only-changes-of';
 
 export interface TypedChange<T> {
   readonly previousValue: T | undefined;
@@ -38,5 +39,75 @@ export class OnChanges$ extends OnDestroy$ implements IOnChanges$ {
 
   ngOnChanges(changes: SimpleChanges): void {
     this.ngOnChangesSubject.next(changes as any);
+  }
+
+  /**
+   * Observe the given property for changes.
+   */
+  input$<P extends keyof this>(property: P): Observable<this[P]> {
+    return input$(this, property);
+  }
+
+
+  /**
+   * Observes the given properties for changes (via OnChanges hook).
+   * Triggers only once for 'simultaneous' changes of multiple inputs. Simultaneous means during one change detection cycle.
+   */
+  inputs$<P extends keyof this>(p1: P): Observable<{ [X in P]: this[X]}>;
+
+  /**
+   * Observes the given properties for changes (via OnChanges hook).
+   * Triggers only once for 'simultaneous' changes of multiple inputs. Simultaneous means during one change detection cycle.
+   */
+  inputs$<P extends keyof this>(p1: P, p2: P): Observable<{ [X in P]: this[X]}>;
+
+  /**
+   * Observes the given properties for changes (via OnChanges hook).
+   * Triggers only once for 'simultaneous' changes of multiple inputs. Simultaneous means during one change detection cycle.
+   */
+  inputs$<P extends keyof this>(p1: P, p2: P, p3: P): Observable<{ [X in P]: this[X]}>;
+
+  /**
+   * Observes the given properties for changes (via OnChanges hook).
+   * Triggers only once for 'simultaneous' changes of multiple inputs. Simultaneous means during one change detection cycle.
+   */
+  inputs$<P extends keyof this>(p1: P, p2: P, p3: P, p4: P): Observable<{ [X in P]: this[X]}>;
+
+  /**
+   * Observes the given properties for changes (via OnChanges hook).
+   * Triggers only once for 'simultaneous' changes of multiple inputs. Simultaneous means during one change detection cycle.
+   */
+  inputs$<P extends keyof this>(p1: P, p2: P, p3: P, p4: P, p5: P): Observable<{ [X in P]: this[X]}>;
+
+  /**
+   * Observes the given properties for changes (via OnChanges hook).
+   * Triggers only once for 'simultaneous' changes of multiple inputs. Simultaneous means during one change detection cycle.
+   */
+  inputs$<P extends keyof this>(p1: P, p2: P, p3: P, p4: P, p5: P, p6: P): Observable<{ [X in P]: this[X]}>;
+
+  /**
+   * Observes the given properties for changes (via OnChanges hook).
+   * Triggers only once for 'simultaneous' changes of multiple inputs. Simultaneous means during one change detection cycle.
+   */
+  inputs$<P extends keyof this>(p1: P, p2: P, p3: P, p4: P, p5: P, p6: P, p7: P): Observable<{ [X in P]: this[X]}>;
+
+  /**
+   * Observes the given properties for changes (via OnChanges hook).
+   * Triggers only once for 'simultaneous' changes of multiple inputs. Simultaneous means during one change detection cycle.
+   */
+  inputs$<P extends keyof this>(p1: P, p2: P, p3: P, p4: P, p5: P, p6: P, p7: P, p8: P): Observable<{ [X in P]: this[X]}>;
+
+  /**
+   * Observes the given properties for changes (via OnChanges hook).
+   * Triggers only once for 'simultaneous' changes of multiple inputs. Simultaneous means during one change detection cycle.
+   */
+  inputs$<P extends keyof this>(p1: P, p2: P, p3: P, p4: P, p5: P, p6: P, p7: P, p8: P, p9: P): Observable<{ [X in P]: this[X]}>;
+
+  // Using var-args parameter breaks refactoring support. Renaming properties will not change the property passed to this method.
+  // I don't think we need more than 9 parameters, so we comment out this option.
+  // inputs$<P extends keyof this>(...properties: P[]): Observable<{ [X in P]: this[X]}>;
+
+  inputs$<P extends keyof this>(...properties: P[]) {
+    return inputs$Impl(this, ...properties);
   }
 }